requirements.txt
beautifulsoup4==4.14.0
psycopg2-binary==2.9.10
requests==2.32.3
Pillow==11.0.0
pika==1.3.2
discord.py==2.4.0
